feat: Accept filenames as CLI args, add pragmas
Add support for optional filenames as CLI args. In the absence of CLI
args, the <./docs/**/*.md> glob will be used.

Add strict and warnings pragmas, use Data::Dumper to view files before
converstion rather than printing inside the loop.

Add a message when the script is complete and a zero exit code.